chore: renaming the milestone form file
