+++ app/mutations/FundingParameter/discardFundingParameterFormChange.ts
@@ -21,7 +21,7 @@ const discardMutation = graphql`
           ...ProjectContactForm_projectRevision
           ...ProjectQuarterlyReportForm_projectRevision
           ...ProjectAnnualReportForm_projectRevision
-          ...ProjectMilestoneReportFormGroup_projectRevision
+          ...ProjectMilestoneReportForm_projectRevision
           ...ProjectFundingAgreementForm_projectRevision
           ...ProjectEmissionIntensityReportForm_projectRevision
         }
